+++ src/onegov.ballot/HISTORY.rst
@@ -8,6 +8,9 @@ Changelog
 - Adds a new domain of influence: region.
   [msom]
 
+- Adds election composites.
+  [msom]
+
 2.0.0 (2018-01-23)
 ~~~~~~~~~~~~~~~~~~~